seperated logic for testing and running the pipeline
